refactor: shift component configs into registry files and simplified
